+++ tests/testthat/test-add-variable.R
@@ -47,12 +47,11 @@ test_that("toVariable handles duplicate factor levels", {
         "Duplicate factor levels given: disambiguating them in translation to Categorical type")
 })
 test_that("categoriesFromLevels parses levels correctly", {
-  expect_identical(
-    categoriesFromLevels(levels(iris$Species)),
-    list(list(id = 1L, name = "setosa", numeric_value = 1L, missing = FALSE),
-         list(id = 2L, name = "versicolor", numeric_value = 2L, missing = FALSE), 
-         list(id = 3L, name = "virginica", numeric_value = 3L, missing = FALSE))
-  )
+    expect_identical(categoriesFromLevels(levels(iris$Species)),
+        list(list(id = 1L, name = "setosa", numeric_value = 1L, missing = FALSE),
+             list(id = 2L, name = "versicolor", numeric_value = 2L, missing = FALSE), 
+             list(id = 3L, name = "virginica", numeric_value = 3L, missing = FALSE))
+      )
 })
 test_that("toVariable parses R Date class", {
     expect_equivalent(toVariable(as.Date(c("2014-12-16", "2014-12-17"))),