+++ NEWS.md
@@ -1,4 +1,6 @@
 ### crunch 1.17.9 (under development)
+* `resetPassword()` function
+* categories are now selectable with names as well as ids.
 
 ### crunch 1.17.8
 Two attempts to fix download issues introduced by 1.17.4: