+++ NEWS.md
@@ -5,6 +5,7 @@
 * Pass a web app URL to `loadDataset()` and it will now load the same dataset in your R session.
 * `webApp()` function to go the other way: open the dataset from your R session in your web browser.
 * `categoriesFromLevels()` is now exported (#77)
+* Add a missing import from the `methods` package so that `Rscript` works (#90)
 
 ### crunch 1.17.8
 Two attempts to fix download issues introduced by 1.17.4: